requirements.txt
hydra-core
langgraph
langchain-community
langchain-anthropic
langchain-groq
langchainhub
langsmith
loguru
pandas
pylint
python-dotenv
pyowm
